Adjust code in Order & Customer components for easier usage
